code/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
google-search-results>=2.4.2
python-dotenv>=1.0.0
matplotlib>=3.7.0
selenium>=4.15.0
